hist refactoring

+++ lib/Integration/gethist.h
@@ -25,7 +25,7 @@ public:
 	hist CloneEmptyBins()const;
 	double Entries()const;
 	point&operator[](int i)const;
-	int count()const;
+	size_t count()const;
 	typedef std::vector<point>::iterator iterator;
 	typedef std::vector<point>::const_iterator const_iterator;
 	iterator begin();
